move converter tests to their own file. Add support for list of mimetypes. Simplified converters as a result of refactoring.
